backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pillow==10.3.0
opencv-python-headless==4.10.0.84
rembg==2.0.57
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0
